requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
audioread==2.1.8
autopep8==1.5
backcall==0.1.0
bleach==3.1.4
bokeh==2.1.1
boto3==1.14.38
botocore==1.17.38
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorcet==2.0.2
cycler==0.10.0
Cython==0.29.16
dask==2.23.0
datashader==0.11.1
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.23.0
docutils==0.15.2
entrypoints==0.3
fsspec==0.8.0
gast==0.3.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-translate==2.0.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
HeapDict==1.0.1
holoviews==1.13.3
httplib2==0.17.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
langdetect==1.0.8
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
msgpack==1.0.0
multidict==4.7.6
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.3
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.0
packaging==20.4
pandarallel==1.4.8
pandas==1.0.3
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.6.2
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.6.0
plotly-express==0.4.0
pooch==1.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyct==0.4.6
PyDrive==1.3.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
PySoundFile==0.9.0.post1
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
pytz==2019.3
pyviz-comms==0.7.6
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sortedcontainers==2.2.2
SoundFile==0.10.3.post1
statsmodels==0.11.1
tblib==1.7.0
telepot==12.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typeguard==2.9.1
typing-extensions==3.7.4.2
umap-learn==0.4.6
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xarray==0.16.0
xmltodict==0.12.0
yarl==1.5.1
zict==2.0.0
zipp==3.1.0
nlpaug==0.0.20